+++ packages/mvc/src/jsx-runtime.ts
@@ -11,6 +11,26 @@ export function jsxs(type: unknown, props: object, key?: unknown): Component.Nod
   return resolved().jsxs(hostType(type), props, key);
 }
 
+/** Dev-transpiled element creation. Falls back to the production runtime when
+ * the host registered no `jsxDEV` (prod-only host under a dev build). */
+export function jsxDEV(
+  type: unknown,
+  props: object,
+  key?: unknown,
+  isStatic?: boolean,
+  source?: object,
+  self?: unknown
+): Component.Node {
+  const runtime = resolved();
+  const resolvedType = hostType(type);
+
+  return runtime.jsxDEV
+    ? runtime.jsxDEV(resolvedType, props, key, isStatic, source, self)
+    : isStatic
+      ? runtime.jsxs(resolvedType, props, key)
+      : runtime.jsx(resolvedType, props, key);
+}
+
 export declare namespace JSX {
   type Element = Component.Node;
   type ElementType =
